feat: require token

+++ src/main.rs
@@ -1,4 +1,5 @@
 mod app;
+mod token;
 
 use std::error::Error;
 use std::{fs, io};
@@ -22,8 +23,8 @@ fn setup_files() -> Result<(), Box<dyn Error>> {
     if let Some(s) = dirs::config_dir() {
         let s = s.join("eng-check");
         fs::create_dir_all(&s)?;
-        File::create(s.join(".token"))?;
-        File::create(s.join("eng-check.conf"))?;
+        if !s.join(".token").exists() { File::create(s.join(".token"))?; }
+        if !s.join("eng-check.conf").exists() { File::create(s.join("eng-check.conf"))?; }
     }
     Ok(())
 }
@@ -38,7 +39,8 @@ async fn main() -> Result<(), Box<dyn Error>> {
     let backend = CrosstermBackend::new(stdout);
     let mut terminal = Terminal::new(backend)?;
 
-    let app = App::new();
+    let token = token::read_or_request().await?;
+    let app = App::new(token);
     app.run(&mut terminal)?;
 
     terminal::disable_raw_mode()?;
